% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{b_helper}
\alias{b_helper}
\title{Beta Helper Function}
\usage{
b_helper(x, y)
}
\arguments{
\item{x}{A real number}

\item{y}{A real number}
}
\value{
A real number equal to sqrt(2)*x/y
}
\description{
This function represents the beta function in my derivation, NOT the beta function defined as an integral.
}
