% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{F_D_asym}
\alias{F_D_asym}
\title{Marginal CDF for Delta Under Asymmetric Exponential Assumption}
\usage{
F_D_asym(d, lam_1, lam_2)
}
\arguments{
\item{d}{The absolute difference between two replicates}

\item{lam_1}{The (positive) parameter for the first replicate's exponential distribution}

\item{lam_2}{The (positive) parameter for the second replicate's exponential distribution}
}
\value{
The marginal probability of observing Delta <= d
}
\description{
This function is the marginal CDF of Delta (which follows an Asymmetric Laplace Distribution) assuming the underlying data X_1 and X_2 follow independent exponential distributions with different parameters.
}
