% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{F_D_sym}
\alias{F_D_sym}
\title{Marginal CDF for Delta Under Symmetric Exponential Assumption}
\usage{
F_D_sym(d, lam)
}
\arguments{
\item{d}{The absolute difference between two replicates}

\item{lam}{The (positive) parameter for the replicates's exponential distributions}
}
\value{
The marginal probability of observing Delta <= d
}
\description{
This function is the marginal CDF of Delta (which follows a Symmetric Laplace Distribution) assuming the underlying data X_1 and X_2 follow independent and identically distributed exponential distributions.
}
