% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{F_exp_asym}
\alias{F_exp_asym}
\title{Joint CDF Under Asymmetric Exponential Assumption}
\usage{
F_exp_asym(d, z, lam_1, lam_2)
}
\arguments{
\item{d}{The aboslute difference between two replicates}

\item{z}{The coefficient of variation between two replicates}

\item{lam_1}{The (positive) parameter for the first replicate's exponential distribution}

\item{lam_2}{The (positive) parameter for the second replicate's exponential distribution}
}
\value{
The joint probability of observing Delta <= d and Z <= z
}
\description{
This function is the joint CDF of Delta and Zeta assuming the underlying data X_1 and X_2 follow independent exponential distributions with different parameters.
}
