% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{F_exp_sym}
\alias{F_exp_sym}
\title{Joint CDF Under Symmetric Exponential Assumption}
\usage{
F_exp_sym(d, z, lam)
}
\arguments{
\item{d}{The aboslute difference between two replicates}

\item{z}{The coefficient of variation between two replicates}

\item{lam}{The (positive) parameter for the replicates's exponential distributions}
}
\value{
The joint probability of observing Delta <= d and Z <= z
}
\description{
This function is the joint CDF of Delta and Zeta assuming the underlying data X_1 and X_2 follow independent and identically distributed exponential distributions.
}
