% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{F_Z_sym}
\alias{F_Z_sym}
\title{Marginal CDF for Zeta Under Symmetric Exponential Assumption}
\usage{
F_Z_sym(z, lam)
}
\arguments{
\item{z}{The coefficient of variation (CV) between two replicates}

\item{lam}{The (positive) parameter for the replicates's exponential distributions}
}
\value{
The marginal probability of observing Zeta <= z
}
\description{
This function is the marginal CDF of Zeta assuming the underlying data X_1 and X_2 follow independent and identically distributed exponential distributions.
}
