% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{g_helper}
\alias{g_helper}
\title{Gamma Helper Function}
\usage{
g_helper(x)
}
\arguments{
\item{x}{A real number (can be vectorized)}
}
\value{
A real number (or vector) equalling (1-x/sqrt(2)) / (1+x/sqrt(2))
}
\description{
This function represents the gamma function in my derivation, NOT the gamma function related to the factorial function.
}
