% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Functions.R
\name{outlier_DZ}
\alias{outlier_DZ}
\title{Original Asymmetric Laplace-Weibull Method for Outlier Detection}
\usage{
outlier_DZ(D, Z, type = "weibull", p_theta = 0.05, p_kappa = 0.05,
  k = 1, t = 0.01, b = 0.01, N_u = 0.01, N_l = 0.01)
}
\arguments{
\item{D}{The absolute difference (Delta) between two vectors of (positive) replicated data: D = X_1 - X_2}

\item{Z}{The coefficient of variation (Zeta) between two vectors of (positive) replicated data: Z = sqrt(2) * abs(X_1 - X_2) / (X_1 + X_2)}

\item{type}{The distribution to which we fit Zeta. By default this value is "weibull", but we can also set it to "lognormal"}

\item{p_theta}{We use the (1-p_theta)*100\% two-sided confidence interval for theta in Delta = X_1 - X_2 + theta to determine if there is a significant translation of the absolute difference Delta. If this interval contains 0, then we set theta = 0. We set p_theta = 0.05 by default}

\item{p_kappa}{We use the (1-p_kappa)*100\% two-sided confidence interval for the asymmetry parameter kappa in the Asymmetric Laplace Distribution to which we fit Delta. If this interval for log(kappa) contains 0, then we set kappa = 0 and use a Symmetric Laplace Distribution for Delta. We set p_kappa = 0.05 by default}

\item{k}{The number of standard deviations about the center (mean) of the Asymmetric Laplace Distribution for Delta that we use to define the "central band." We set k = 1 by default}

\item{t}{We remove the top t*100\% of points when fitting the coefficient of variation Zeta values to a Weibull or lognormal distribution, so that these points do not cause problems when fitting the distribution. We set t = 0.01 by default}

\item{b}{We remove the bottom b*100\% of points when fitting the coefficient of variation Zeta values to a Weibull or lognormal distribution, so that these points do not cause problems when fitting the distribution. We set b = 0.01 by default}

\item{N_u}{After we fit the Zeta values for the correct subset of points to a Weibull or lognormal distribution, we return as outliers all the points whose Zeta values are greater than the level above which we would not expect to find N_u*100\% of the points. This procedure is how the underlying function getOutliersI in the package extremevalues works. We set N_u = 1 by default}

\item{N_l}{After we fit the Zeta values for the correct subset of points to a Weibull or lognormal distribution, we return as outliers all the points whose Zeta values are less than the level below which we would not expect to find N_l*100\% of the points. This procedure is how the underlying function getOutliersI in the package extremevalues works. We set N_l = 1 by default}
}
\value{
A numerical vector of equal length to the input D and Z vectors, with values 0 for points in the central band, 1 for points in the "wings" with small Zeta, and 2 for outliers
}
\description{
This function implements the original method for outlier detection among replicated data. It first fits the aboslute difference Delta between two replicates to an Asymmetric Laplace Distribution using MLE. Then among the points outside of some central band, we fit their values for the coefficient of variation Zeta to a Weibull (or lognormal) distribution (also using MLE). The points from this set with significantly large Zeta values are outliers.
}
\examples{
# Assume X_1 and X_2 are positive data vectors of the same length. These are the replicates
df <- data.frame(X_1=X_1, X_2=X_2)
df$D <- df$X_1 - df$X_2
df$Z <- sqrt(2) * abs(df$X_1 - df$X_2) / (df$X_1 + df$X_2)
df$outlier_status <- outlier_DZ(df$D, df$Z)
}
