% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{q_exp_asym}
\alias{q_exp_asym}
\title{Outlier Probability Under Asymmetric Exponential Assumption (Using Joint Method)}
\usage{
q_exp_asym(d, z, lambda_1, lambda_2)
}
\arguments{
\item{d}{The absolute difference between two replicates}

\item{z}{The coefficient of variation (CV) between two replicates}

\item{lambda_1}{The (positive) parameter for the first replicate's exponential distribution}

\item{lambda_2}{The (positive) parameter for the second replicate's exponential distribution}
}
\value{
The joint probability of observing Delta <= d and Z >= z (if d < 0) or Delta >= d and Z >= z (if d >= 0)
}
\description{
This function provides the outlier probability of a pair of replicates described by (Delta=d, Zeta=z) using the joint method assuming the underlying data X_1 and X_2 follow independent exponential distributions with different parameters.
}
