% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{q_exp_sym}
\alias{q_exp_sym}
\title{Outlier Probability Under Symmetric Exponential Assumption (Using Joint Method)}
\usage{
q_exp_sym(d, z, lambda)
}
\arguments{
\item{d}{The absolute difference between two replicates}

\item{z}{The coefficient of variation (CV) between two replicates}

\item{lambda}{The (positive) parameter for the replicates's exponential distributions}
}
\value{
The joint probability of observing Delta <= d and Z >= z (if d < 0) or Delta >= d and Z >= z (if d >= 0)
}
\description{
This function provides the outlier probability of a pair of replicates described by (Delta=d, Zeta=z) using the joint method assuming the underlying data X_1 and X_2 follow independent and identically distributed exponential distributions.
}
