% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{q_gg_Z}
\alias{q_gg_Z}
\title{Outlier Probability for CV Under GG Assumption (Using Marginal Method)}
\usage{
q_gg_Z(z, alpha_1, beta_1, c_1, alpha_2, beta_2, c_2)
}
\arguments{
\item{z}{The coefficient of variation (CV) between two replicates}

\item{alpha_1}{The parameter alpha > 0 in the GG distribution for X_1}

\item{beta_1}{The parameter beta > 0 in the GG distribution for X_1}

\item{c_1}{The parameter c > 0 in the GG distribution for X_1}

\item{alpha_2}{The parameter alpha > 0 in the GG distribution for X_2}

\item{beta_2}{The parameter beta > 0 in the GG distribution for X_2}

\item{c_2}{The parameter c > 0 in the GG distribution for X_2}
}
\value{
The marginal probability of observing Zeta >= z
}
\description{
This function provides the outlier probability of a pair of replicates described by Zeta = z using the marginal method assuming the underlying data X_1 and X_2 follow independent (and possibly identically distributed) GG distributions. We use the parameterization used by the R package flexsurv, given by Kotz and Johnson (1970). We numerically integrate the PDF given by the function f_gg using the function adaptIntegrate in the package cubature.
}
